added parse_sec_cfg
