+++ asdit/datasets/collators.py
@@ -6,7 +6,7 @@ import numpy as np
 import torch
 
 from asdit.utils.config_class import LabelInfo
-from asdit.utils.dcase_utils import dcase2sec, get_dcase_info, get_label_dict
+from asdit.utils.dcase_utils import get_dcase_info, get_label_dict
 
 logger = logging.getLogger(__name__)
 
@@ -61,10 +61,9 @@ class BasicCollator(object):
             logging.info(
                 "`wave` in batch will be ignored. If you need it, set `sec` parameter."
             )
+        elif isinstance(sec, str):
+            raise ValueError(f"Unexpected sec: {sec}")
         else:
-            if isinstance(sec, str):
-                sec = dcase2sec(sec)
-            assert type(sec) is float or type(sec) is int
             self.crop_len = int(sr * sec)
 
         self.need_wave: bool = self.crop_len != "none"
