+++ asdit/models/multiresnet/multiresnet.py
@@ -3,8 +3,6 @@ from typing import List
 import torch
 from torch import nn
 
-from asdit.utils.dcase_utils import dcase2sec
-
 from .fft_encoder import FFTEncoderLayer
 from .stft_encoder import STFTEncoderLayer
 
@@ -12,7 +10,7 @@ from .stft_encoder import STFTEncoderLayer
 class MultiResNet(nn.Module):
     def __init__(
         self,
-        sec: float | str,
+        sec: float,
         sr: int,
         use_fft: bool,
         stft_cfg_list: List[dict],
@@ -21,8 +19,6 @@ class MultiResNet(nn.Module):
         resnet_additional_layer: str = "SEBlock",
     ):
         super().__init__()
-        if isinstance(sec, str):
-            sec = dcase2sec(sec)
 
         if use_fft:
             self.fft_layer = FFTEncoderLayer(
