+++ asdit/utils/dcase_utils/__init__.py
@@ -1,6 +1,6 @@
-from .dcase2sec import dcase2sec, dcase2sec_dict
 from .dcase_idx import get_dcase_idx, get_domain_idx
 from .dcase_info import get_dcase_info
 from .idx_util import get_label_dict
 from .infolist import INFOLIST
 from .machines import MACHINE_DICT
+from .parse_sec import parse_sec_cfg
